homeController test class split in 2
